Rename safe_run_dkg -> run_dkg, status -> structures
