+++ web/models/schain.py
@@ -24,7 +24,7 @@ from datetime import datetime
 from peewee import (CharField, DateTimeField,
                     IntegrityError, IntegerField, BooleanField)
 
-from core.schains.dkg.status import DKGStatus
+from core.schains.dkg.structures import DKGStatus
 from web.models.base import BaseModel
 
 logger = logging.getLogger(__name__)
